refactor: shared UIs 각 컴포넌트별로 파일 분리
